// src/app/core/models/quiz.model.ts
export interface Choice {
    choice_id: number;
    question_id: number;
    choice: string;
    recommend: boolean;
  }
  
  export interface Question {
    quiz_id: number;
    cat_name: string;
    questions: string;
    marks: number;
    choices: Choice[];
  }
  
  export interface Pagination {
    current_page: number;
    page_size: number;
    total_items: number;
    total_pages: number;
  }
  
  export interface QuizResponse {
    status: string;
    message: string;
    data: {
      data: Question[];
      pagination: Pagination;
    };
  }
  
  export interface UserAnswer {
    questionId: number;
    selectedChoiceId: number;
  }